requirements/base.txt
# Core Django and API dependencies
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Cache and Redis
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Celery for async tasks
celery==5.3.6
celery[redis]==5.3.6
django-celery-beat==2.6.0

# Payment Gateways
razorpay==1.4.1
stripe==8.2.0

# AI/LLM
anthropic==0.18.1
openai==1.12.0

# Email
django-anymail[mailgun]==10.2

# SMS
twilio==8.13.0

# API Documentation
drf-spectacular==0.27.1

# Environment variables
python-decouple==3.8
python-dotenv==1.0.1

# Date/Time utilities
python-dateutil==2.8.2
pytz==2024.1

# HTTP requests
requests==2.31.0

# Security
cryptography==42.0.5

# Monitoring and logging
sentry-sdk==1.40.6

# Utils
Pillow==10.2.0
